+++ cli.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"context"
 	"errors"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"strings"
@@ -15,60 +15,20 @@ import (
 	yaml "gopkg.in/yaml.v2"
 )
 
-type Labeler interface {
-	GetLabel(ctx context.Context, owner string, repo string, name string) (*github.Label, *github.Response, error)
-	EditLabel(ctx context.Context, owner string, repo string, name string, label *github.Label) (*github.Label, *github.Response, error)
-	CreateLabel(ctx context.Context, owner string, repo string, label *github.Label) (*github.Label, *github.Response, error)
-	ListLabels(ctx context.Context, owner string, repo string, opt *github.ListOptions) ([]*github.Label, *github.Response, error)
-	DeleteLabel(ctx context.Context, owner string, repo string, name string) (*github.Response, error)
+type CLI struct {
+	Stdout io.Writer
+	Stderr io.Writer
+	Option Option
+
+	GitHub *githubClient
+	Config Manifest
 }
 
-type githubClientImpl struct {
-	ghClient *github.Client
-}
-
-func (l githubClientImpl) GetLabel(ctx context.Context, owner string, repo string, name string) (*github.Label, *github.Response, error) {
-	return l.ghClient.Issues.GetLabel(ctx, owner, repo, name)
-}
-
-func (l githubClientImpl) EditLabel(ctx context.Context, owner string, repo string, name string, label *github.Label) (*github.Label, *github.Response, error) {
-	return l.ghClient.Issues.EditLabel(ctx, owner, repo, name, label)
-}
-
-func (l githubClientImpl) CreateLabel(ctx context.Context, owner string, repo string, label *github.Label) (*github.Label, *github.Response, error) {
-	return l.ghClient.Issues.CreateLabel(ctx, owner, repo, label)
-}
-
-func (l githubClientImpl) ListLabels(ctx context.Context, owner string, repo string, opt *github.ListOptions) ([]*github.Label, *github.Response, error) {
-	return l.ghClient.Issues.ListLabels(ctx, owner, repo, opt)
-}
-
-func (l githubClientImpl) DeleteLabel(ctx context.Context, owner string, repo string, name string) (*github.Response, error) {
-	return l.ghClient.Issues.DeleteLabel(ctx, owner, repo, name)
-}
-
-type githubClientDryRun struct {
-	ghClient *github.Client
-}
-
-func (l githubClientDryRun) GetLabel(ctx context.Context, owner string, repo string, name string) (*github.Label, *github.Response, error) {
-	return l.ghClient.Issues.GetLabel(ctx, owner, repo, name)
-}
-
-func (l githubClientDryRun) EditLabel(ctx context.Context, owner string, repo string, name string, label *github.Label) (*github.Label, *github.Response, error) {
-	return nil, nil, nil
-}
-
-func (l githubClientDryRun) CreateLabel(ctx context.Context, owner string, repo string, label *github.Label) (*github.Label, *github.Response, error) {
-	return nil, nil, nil
-}
-
-func (l githubClientDryRun) ListLabels(ctx context.Context, owner string, repo string, opt *github.ListOptions) ([]*github.Label, *github.Response, error) {
-	return l.ghClient.Issues.ListLabels(ctx, owner, repo, opt)
-}
-
-func (l githubClientDryRun) DeleteLabel(ctx context.Context, owner string, repo string, name string) (*github.Response, error) {
-	return nil, nil
+type Option struct {
+	DryRun  bool   `long:"dry-run" description:"Just dry run"`
+	Config  string `short:"c" long:"config" description:"Path to YAML file that labels are defined" default:"labels.yaml"`
+	Import  bool   `long:"import" description:"Import existing labels if enabled"`
+	Version bool   `long:"version" description:"Show version"`
 }
 
 type githubClient struct {
